structural standardization
