+++ examples/advanced/encapsulate-use-fetch/api/dto/globle.dto.ts
