+++ examples/advanced/encapsulate-use-fetch/api/hackernews.ts
@@ -1,0 +1,8 @@
+import { FeedsDataT, FeedsModel } from './model/feeds.model';
+import { request } from './util/request'
+
+enum Api {
+    Feeds = '/api/hn/feeds',
+}
+
+export const FeedsApi = (params: FeedsModel) => request<FeedsDataT>( Api.Feeds, { method: 'GET', params: params } );
