+++ examples/advanced/encapsulate-use-fetch/api/utils/http/request.ts
